rename benchmark->multi
